+++ GherkinSpec.TestAdapter.CustomBindingAttributesUnitTests/CustomAttributes/CustomStepsAttribute.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace GherkinSpec.TestAdapter.CustomBindingAttributesUnitTests.CustomAttributes;
+
+[AttributeUsage(AttributeTargets.Class)]
+public class CustomStepsAttribute : Attribute
+{
+}