+++ GherkinSpec.TestAdapter.CustomBindingAttributesUnitTests/CustomAttributes/CustomThenAttribute.cs
@@ -1,0 +1,15 @@
+﻿using System;
+
+namespace GherkinSpec.TestAdapter.CustomBindingAttributesUnitTests.CustomAttributes;
+
+[AttributeUsage(AttributeTargets.Method)]
+public class CustomThenAttribute : Attribute, ICustomStepAttribute
+{
+    public CustomThenAttribute(
+        string matchRegex)
+    {
+        MatchRegex = matchRegex;
+    }
+
+    public string MatchRegex { get; }
+}