+++ GherkinSpec.TestAdapter.CustomBindingAttributesUnitTests/CustomAttributes/ICustomStepAttribute.cs
@@ -1,0 +1,6 @@
+﻿namespace GherkinSpec.TestAdapter.CustomBindingAttributesUnitTests.CustomAttributes;
+
+public interface ICustomStepAttribute
+{
+    public string MatchRegex { get; }
+}