+++ GherkinSpec.TestAdapter.UnitTests.ReferencedAssembly/NonStaticStepsClass.cs
@@ -1,0 +1,9 @@
+﻿using GherkinSpec.TestModel;
+
+namespace GherkinSpec.TestAdapter.UnitTests.ReferencedAssembly
+{
+    [Steps]
+    public class NonStaticStepsClass
+    {
+    }
+}