+++ GherkinSpec.TestModel.UnitTests/ServiceCollectionExtensionsShould.cs
@@ -1,7 +1,8 @@
-﻿using GherkinSpec.TestModel.UnitTests.Samples;
+﻿using System.Linq;
+using GherkinSpec.TestAdapter.UnitTests.ReferencedAssembly;
+using GherkinSpec.TestModel.UnitTests.Samples;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System.Linq;
 
 namespace GherkinSpec.TestModel.UnitTests
 {
@@ -30,18 +31,27 @@ namespace GherkinSpec.TestModel.UnitTests
         public void RegisterAllNonStaticStepsClassesInSpecifiedAssembly()
         {
             var services = new ServiceCollection();
-            services.AddAllStepsClassesAsScoped(typeof(ReferencedAssembly.StaticStepsClass).Assembly);
+            services.AddAllStepsClassesAsScoped(typeof(StaticStepsClass).Assembly);
 
-            Assert.AreEqual(typeof(ReferencedAssembly.NonStaticStepsClass), services.Single().ServiceType);
+            var expectedNonStaticStepsClasses = new[]
+            {
+                typeof(NonStaticStepsClass),
+                typeof(ReferencedAssemblyHookSamples),
+                typeof(ReferencedAssemblyStepBindingSamples)
+            };
+            
+            CollectionAssert.AreEquivalent(
+                expectedNonStaticStepsClasses,
+                services.Select(s => s.ServiceType).ToList());
         }
 
         [TestMethod]
         public void DoesNotRegisterStaticStepsClassesInSpecifiedAssembly()
         {
             var services = new ServiceCollection();
-            services.AddAllStepsClassesAsScoped(typeof(ReferencedAssembly.StaticStepsClass).Assembly);
+            services.AddAllStepsClassesAsScoped(typeof(StaticStepsClass).Assembly);
 
-            Assert.IsFalse(services.Any(descriptor => descriptor.ServiceType == typeof(ReferencedAssembly.StaticStepsClass)));
+            Assert.IsFalse(services.Any(descriptor => descriptor.ServiceType == typeof(StaticStepsClass)));
         }
     }
 }