+++ GherkinSpec.TestModel/IStepAttributeMatchExpressionSelector.cs
@@ -1,0 +1,6 @@
+namespace GherkinSpec.TestModel;
+
+public interface IStepAttributeMatchExpressionSelector
+{
+    string SelectMatchExpression(object attribute);
+}