+++ Changelog.md
@@ -2,6 +2,10 @@
 
 ## 3.0
 
+### Added
+
+* Support for specified binding attribute types (see the [`Custom Binding Attribute Types documentation`](./docs/CustomBindingAttributeTypes.md)).
+
 ### Changed
 
 * Updated project target frameworks (to `netstandard2.1` for published packages & `net6.0` for unit test projects).